EVPRO-1393

+++ evolution/src/main/java/com/evolving/nglm/evolution/GUIManagerBaseManagement.java
@@ -1818,7 +1818,14 @@ public class GUIManagerBaseManagement extends GUIManager
         UCGRuleIDs.add(UCGRuleID);
         GUIManagedObject ucgRule = ucgRuleService.getStoredUCGRule(UCGRuleID);
         if (ucgRule != null && (force || !ucgRule.getReadOnly()))
-          singleIDresponseCode = "ok";
+          if(ucgRule.getActive())
+          {
+            singleIDresponseCode = "deletingActiveNotAllowed";
+          }
+        else
+          {
+            singleIDresponseCode = "ok";
+        }
         else if (ucgRule != null)
           singleIDresponseCode = "failedReadOnly";
         else singleIDresponseCode = "ucgRuleNotFound";
@@ -1855,8 +1862,10 @@ public class GUIManagerBaseManagement extends GUIManager
       {
 
         GUIManagedObject ucgRule = UCGRules.get(i);
-
-        ucgRuleService.removeUCGRule(ucgRule.getGUIManagedObjectID(), userID, tenantID);
+        if(!ucgRule.getActive())
+        {
+          ucgRuleService.removeUCGRule(ucgRule.getGUIManagedObjectID(), userID, tenantID);
+        }
       }
 
     /*****************************************
